django/requirements.txt
django==5.0
djangorestframework==3.15
psycopg2-binary==2.9.9
PyJWT==2.8.0
django-cors-headers==4.3.1
python-decouple==3.8
